uploaded files from computer usbwebserver
I have several files that I will not use all of the functionality that I have made also there are sure to be some bugs and incomplete and not implemented code. which I will go back and delete when I know it does not serve any useful function. I wrote all of this code over the years without the use of git or GitHub.   

Thank you
Zac Owens
Zakotech